+++ src/font.h
@@ -4,10 +4,11 @@
 
 extern unsigned char font[256*16];
 
-void draw_char(int x, int y, unsigned char ch);
-void draw_string(int x, int y, int w, const char * str, int spacing);
-void draw_string_centered(int x, int y, int w,
-		const char * str, int spacing);
+void draw_char(float x, float y, unsigned char ch);
+void draw_string(float x, float y, float w,
+		const char * str, float spacing);
+void draw_string_centered(float x, float y, float w,
+		const char * str, float spacing);
 
 #endif
